feat: support FlightSQL
This commit brings support to the influxdb3 server to handle gRPC
requests, i.e., FlightSQL requests.

The primary server will route gRPC reqeusts to an underlying
FlightService, while other requests will still go through to the
standard REST HTTP server.

A test was added to check that it was actually working, along with
some common test helpers to spawn a server for testing.