test: add test to check InfluxQL functionality
Some additional test cases were added to test some of the InfluxQL
features:

- SHOW MEASUREMENTS
- SHOW FIELD KEYS
- SHOW TAG KEYS

Also, a case was added for SHOW TAG VALUES [...] but that query looks
to not be working. This is not expected.

Other queries that would not work would be those that are performed
above the DATABASE level, e.g., SHOW DATABASES, or queries that use
the ON <db_name> syntax.