feat: implement loader for persisted state
This implements a loader for the write buffer. It loads the catalog and the buffer from the WAL.

Move Persister errors into their own type now that the write buffer load could return errors from the persister.

This doesn't yet rotate segments or trigger persistence of newly closed segments, which will be addressed in a future PR.