feat: support InfluxQL queries containing db
This extends Edge's InfluxQL capability to support InfluxQL
queries that operate on a particular database. Generally, these
are queries that use an ON <db_name> clause, but database names
can also appear in the FROM clause in a SELECT statement.

This brings changes to core crates, which will need to be upstreamed
if they pass here.

This also adds support for the SHOW DATABASES query.

Tests were added in the e2e suite for the query_influxql API