refactor: move InfluxQL rewrite logic to new crate
This moves the logic that was added to the iox_query_influxql crate
into a separate crate to mitigate the delta with IOx and what will
become a core crate in the near future.

The new crate, iox_query_influxql_rewrite is named to imply that it
should eventually merge with the former.