feat: Change Bearer Auth Token to use random bits
This changes the 'influxdb3 create token' command so that it will just
automatically generate a completely random token prepended with
'apiv3_' that is then fed into a Sha512 algorithm instead of Sha256. The
user can no longer pass in a token to be turned into the proper output.

This also changes the server code to handle the change to Sha512 as
well.

Closes #24704