feat: implement Authorizer and use to auth REST and gRPC
This adds two new Authorizer implementations to Edge: Default and
AllOrNothing, which will provide the two auth options for Edge.

A builder was added to the Server in this commit, as part of an
attempt to get the server creation to be more modular.

Both gRPC requests and HTTP REST request will be authorized by
the same Authorizer implementation.