test: add test for authorized gRPC
A new end-to-end test, auth_grpc, was added to check that
authorization is working with the influxdb3 Flight service.

The SHA512 digest action was moved into the Authorizer
implementation so that both HTTP and gRPC auth worked.

The test fixtures had some updates to ensure that auth tokens
are set correctly.