feat: add _series_id to tables on write
New _series_id column is added to tables written to; this stores
a 32 byte SHA256 hash of the tag set of a line of Line Protocol.
The tag set is checked for sort order, then sorted if not already,
before producing the hash.

Unit tests were added to check hashing and sorting functions work,
although, these changes broke existing tests that need to now be
fixed.