feat: Create initial pro specific changes (#4)
This commit does a few things so show what working in the pro repo can
look like from a code perspective as well as minimizing code conflict
*between* pro and edge. It also sets us up to work on pro specific http
endpoints as well as setting up the skeleton of a compactor.

Neither of these changes does anythings particularly interesting. For
the pro http endpoint we just echo back what gets posted and for the
compactor we spawn a task that just sleeps in a continuous loop while
having access to the memory buffer and persister on startup. Overtime
the compactor can be expanded to work on actually compacting parquet
files and as for the pro endpoints we can add things as needed, such as
OAuth2 etc.