feat: start on series key implementation in monolith
This commit points at the series key branch on influxdb3_core, and refac-
tors the code to make use of it throughout.

TableDefinitions in the catalog are updated to support the series key
column type, and will cause panics when de/serializing table snapshots
that contain cross-contaminated column types, i.e., that have both tags
and series key columns.

This does not implement the v3 write API or any of the stack that
performs the writes to use these new TableDefinition features.

Several tests are broken by this commit, so may be fixed in a future
commit, or with ammendments, but this is a stopping point before pursuing
further changes.