feat: support addition of newly written columns to last cache (#25125)
* feat: support last caches that can add new fields

* feat: support new values in last cache

Support the addition of new fields to the last cache, for caches that do
not have a specified set of value columns.

A test was added along with the changes.

* chore: clippy

* docs: add comments throughout new last cache code

* fix: last cache schema merging when new fields added

* refactor: use outer schema for RecordBatch production

Enabling addition of new fields to a last cache made the insertion order
guarantee of the IndexMap break down. It could not be relied upon anymore
so this commit removes reference to that fact, despite still using the
IndexMap type, and strips out the schema from the inner LastCacheStore
type of the LastCache.

Now, the outer LastCache schema is relied on for producing RecordBatches,
which requires a lookup to the inner LastCacheStore's HashMap for each
field in the schema. This may not be as convenient as iterating over the
map as before, but trying to manage the disparate schema, and maintaining
the map ordering was making the code too complicated. This seems like a
reasonable compromise for now, until we see the need to optimize.

The IndexMap is still used for its fast iteration and lookup
characteristics.

The test that checks for new field ordering behaviour was modified to be
correct.

* refactor: use has set instead of scanning entire row on each push

Some renaming of variables was done to clarify meaning as well.